Fixed calendar size issue, broken blog images
The calendar margin was 200px, I increased it to 400px.
I needed to do a full site render to push previous edits through to all pages, but for some reason a few of the old blog images started throwing errors when I did the full site render. I replaced those images, and that seemed to fix it.